collecting_data/ver_two/requirements.txt
numpy
opencv-python
torch 
torchvision
scikit-learn
scikit-image
six
requests
pymilvus==1.0.1
tqdm
SQLAlchemy==1.3.15
psycopg2-binary==2.9.3